Completed the Riverpod Implementation
