+++ lib/upi/provider/notifiers.dart
@@ -34,22 +34,3 @@ class ShowArchivedNotifier extends StateNotifier<bool> {
     state = !state;
   }
 }
-
-// StateNotifier for managing accounts
-class AccountsNotifier extends StateNotifier<List<Accounts>> {
-  AccountsNotifier() : super([]);
-
-  // Add a method to load accounts, perhaps from a database
-  void loadAccounts(List<Accounts> accounts) {
-    state = accounts;
-  }
-
-// Add other methods for modifying the accounts list if needed
-}
-
-// StateNotifier for products
-class ProductNotifier extends StateNotifier<List<Product>> {
-  ProductNotifier() : super([]);
-
-// Add methods to fetch and update products if needed
-}
