+++ lib/upi/screens/transaction_screen.dart
@@ -22,349 +22,6 @@ final productsBoxProvider = FutureProvider<Box<Product>>((ref) async {
   return await Hive.openBox<Product>('products');
 });
 
-// class TransactionScreen extends StatefulWidget {
-//   final Accounts? account;
-//
-//   const TransactionScreen({super.key, this.account});
-//
-//   @override
-//   // ignore: library_private_types_in_public_api
-//   _TransactionScreenState createState() => _TransactionScreenState();
-// }
-//
-// class _TransactionScreenState extends State<TransactionScreen> {
-//   String _selectedStatus = 'all';
-//   String _searchQuery = '';
-//   DateTime? _startDate;
-//   DateTime? _endDate;
-//   final LocalAuthentication auth = LocalAuthentication();
-//
-//   Timer? _inactivityTimer;
-//   static const int _inactivityDuration = 10 * 60; // 10 minutes in seconds
-//
-//   @override
-//   void initState() {
-//     super.initState();
-//     _startInactivityTimer();
-//   }
-//
-//   @override
-//   void dispose() {
-//     _inactivityTimer?.cancel();
-//     super.dispose();
-//   }
-//
-//   // Method to start the inactivity timer
-//   void _startInactivityTimer() {
-//     _inactivityTimer?.cancel(); // Cancel any existing timer
-//     _inactivityTimer = Timer(Duration(seconds: _inactivityDuration), () async {
-//       bool isAuthenticated = await authenticate();
-//       if (!isAuthenticated) {
-//         // Handle failed authentication (e.g., lock the screen or navigate to a login screen)
-//         print('Authentication failed after inactivity.');
-//       } else {
-//         print('User authenticated successfully after inactivity.');
-//       }
-//     });
-//   }
-//
-//   // Reset the inactivity timer on user activity
-//   void _resetInactivityTimer() {
-//     _startInactivityTimer();
-//   }
-//
-//   Future<bool> authenticate() async {
-//     return true;
-//     try {
-//       bool canCheckBiometrics = await auth.canCheckBiometrics;
-//       bool isAuthenticated = false;
-//
-//       if (canCheckBiometrics) {
-//         isAuthenticated = await auth.authenticate(
-//           localizedReason: 'Please authenticate to access the app',
-//           options: const AuthenticationOptions(biometricOnly: true),
-//         );
-//       } else {}
-//
-//       return isAuthenticated;
-//     } catch (e) {
-//       return false;
-//     }
-//   }
-//
-//   void _showDateRangePicker(BuildContext context) {
-//     showModalBottomSheet(
-//       context: context,
-//       builder: (context) {
-//         DateTime? tempStartDate = _startDate;
-//         DateTime? tempEndDate = _endDate;
-//         Map<DateTime, List<Order>> ordersPerDay = {};
-//
-//         final orders = Hive.box<Order>('orders').values;
-//         print('Total orders in Hive box: ${orders.length}');
-//         for (var order in orders) {
-//           print('Order timestamp: ${order.timestamp}');
-//           final orderDate = DateTime(
-//               order.timestamp.year, order.timestamp.month, order.timestamp.day);
-//           print('Normalized order date: $orderDate');
-//           if (ordersPerDay.containsKey(orderDate)) {
-//             ordersPerDay[orderDate]!.add(order);
-//           } else {
-//             ordersPerDay[orderDate] = [order];
-//           }
-//         }
-//
-//         return StatefulBuilder(
-//           builder: (BuildContext context, StateSetter setModalState) {
-//             return Container(
-//               padding: const EdgeInsets.all(16.0),
-//               child: Column(
-//                 mainAxisSize: MainAxisSize.min,
-//                 children: [
-//                   TableCalendar(
-//                     focusedDay: DateTime.now(),
-//                     firstDay: DateTime(2000),
-//                     lastDay: DateTime(2100),
-//                     rangeSelectionMode: RangeSelectionMode.toggledOn,
-//                     selectedDayPredicate: (day) =>
-//                     isSameDay(day, tempStartDate) ||
-//                         isSameDay(day, tempEndDate),
-//                     rangeStartDay: tempStartDate,
-//                     rangeEndDay: tempEndDate,
-//                     onRangeSelected: (start, end, focusedDay) {
-//                       setModalState(() {
-//                         tempStartDate = start;
-//                         tempEndDate = end;
-//                       });
-//                     },
-//                     calendarStyle: const CalendarStyle(
-//                       outsideDaysVisible: false,
-//                     ),
-//                     eventLoader: (day) {
-//                       final normalizedDay =
-//                       DateTime(day.year, day.month, day.day);
-//                       return ordersPerDay[normalizedDay] ?? [];
-//                     },
-//                     onDaySelected: (selectedDay, focusedDay) {
-//                       setModalState(() {
-//                         if (tempStartDate == null ||
-//                             (tempStartDate != null && tempEndDate != null)) {
-//                           tempStartDate = selectedDay;
-//                           tempEndDate = null;
-//                         } else if (tempEndDate == null) {
-//                           tempEndDate = selectedDay;
-//                         }
-//                       });
-//                     },
-//                     calendarBuilders: CalendarBuilders(
-//                       markerBuilder: (context, day, events) {
-//                         if (events.isNotEmpty) {
-//                           return Positioned(
-//                             bottom: 0,
-//                             right: 2.0,
-//                             child: Container(
-//                               decoration: const BoxDecoration(
-//                                 color: Colors.blue,
-//                                 shape: BoxShape.circle,
-//                               ),
-//                               padding: const EdgeInsets.all(6.0),
-//                               child: Center(
-//                                 child: Text(
-//                                   '${events.length}',
-//                                   style: const TextStyle(
-//                                     color: Colors.white,
-//                                     fontSize: 14.0,
-//                                   ),
-//                                 ),
-//                               ),
-//                             ),
-//                           );
-//                         } else {
-//                           return const SizedBox();
-//                         }
-//                       },
-//                     ),
-//                   ),
-//                   const SizedBox(height: 20),
-//                   ElevatedButton(
-//                     onPressed: () {
-//                       setState(() {
-//                         _startDate = tempStartDate;
-//                         _endDate = tempEndDate;
-//                       });
-//                       Navigator.pop(context);
-//                     },
-//                     child: const Text('Apply Date Filter'),
-//                   ),
-//                 ],
-//               ),
-//             );
-//           },
-//         );
-//       },
-//     );
-//   }
-//
-//   @override
-//   Widget build(BuildContext context) {
-//     return FutureBuilder<bool>(
-//       future: authenticate(),
-//       builder: (context, snapshot) {
-//         if (snapshot.connectionState == ConnectionState.waiting) {
-//           return const Scaffold(
-//             body: Center(child: CircularProgressIndicator()),
-//           );
-//         } else if (snapshot.hasError || !snapshot.data!) {
-//           return const Scaffold(
-//             body: Center(child: Text('Authentication failed')),
-//           );
-//         } else {
-//           return customBody(context);
-//         }
-//       },
-//     );
-//   }
-//
-//   Scaffold customBody(BuildContext context) {
-//     return Scaffold(
-//       appBar: AppBar(
-//         title: const Text('Recent Transactions'),
-//       ),
-//       body: Column(
-//         children: [
-//           Padding(
-//             padding: const EdgeInsets.all(16.0),
-//             child: TextField(
-//               onChanged: (value) {
-//                 setState(() {
-//                   _searchQuery = value.toLowerCase();
-//                 });
-//               },
-//               decoration: const InputDecoration(
-//                 hintText: 'Search by product...',
-//                 hintStyle: TextStyle(color: Colors.black54),
-//                 border: OutlineInputBorder(),
-//               ),
-//             ),
-//           ),
-//
-//           // Search, Filter, and Date Picker Section
-//           Padding(
-//             padding: const EdgeInsets.all(16.0),
-//             child: Row(
-//               children: [
-//                 // Search Bar
-//
-//                 const SizedBox(width: 16),
-//                 // Status Filter Dropdown
-//                 DropdownButton<String>(
-//                   value: _selectedStatus,
-//                   items: <String>['all', 'pending', 'completed']
-//                       .map((String value) {
-//                     return DropdownMenuItem<String>(
-//                       value: value,
-//                       child: Text(value),
-//                     );
-//                   }).toList(),
-//                   onChanged: (String? newValue) {
-//                     setState(() {
-//                       _selectedStatus = newValue ?? 'all';
-//                     });
-//                   },
-//                   icon: const Icon(Icons.filter_list),
-//                 ),
-//                 const SizedBox(width: 16),
-//                 // Date Range Picker Button
-//                 IconButton(
-//                   icon: const Icon(Icons.date_range),
-//                   onPressed: () {
-//                     _showDateRangePicker(context);
-//                   },
-//                 ),
-//                 if (_startDate != null && _endDate != null)
-//                   Padding(
-//                     padding: const EdgeInsets.symmetric(horizontal: 1.0),
-//                     child: Row(
-//                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
-//                       children: [
-//                         Text(
-//                           '${DateFormat('d MMM, yy').format(_startDate!)} -'
-//                               '${DateFormat('d MMM, yy').format(_endDate!)}',
-//                           style: const TextStyle(
-//                               fontSize: 14, fontWeight: FontWeight.bold),
-//                         ),
-//                         IconButton(
-//                           onPressed: () {
-//                             setState(() {
-//                               _startDate = null;
-//                               _endDate = null;
-//                             });
-//                           },
-//                           icon: const Icon(Icons.close),
-//                         ),
-//                       ],
-//                     ),
-//                   ),
-//               ],
-//             ),
-//           ),
-//           // Display Selected Date Range and Clear Button
-//
-//           // Orders List
-//           Expanded(
-//             child: ValueListenableBuilder(
-//               valueListenable: Hive.box<Order>('orders').listenable(),
-//               builder: (context, Box<Order> box, _) {
-//                 final productBox = Hive.box<Product>('products');
-//                 final orders = box.values.where((order) {
-//                   final matchesStatus = _selectedStatus == 'all' ||
-//                       order.status == _selectedStatus;
-//
-//                   final matchesAccount = widget.account == null ||
-//                       order.products.keys.any((productId) {
-//                         return widget.account!.productIds.contains(productId);
-//                       });
-//
-//                   final matchesSearch = _searchQuery.isEmpty ||
-//                       order.products.keys.any((productId) {
-//                         final product = productBox.get(productId);
-//                         return product != null &&
-//                             product.name.toLowerCase().contains(_searchQuery);
-//                       });
-//
-//                   final matchesDate = (_startDate == null ||
-//                       order.timestamp.isAfter(_startDate!)) &&
-//                       (_endDate == null || order.timestamp.isBefore(_endDate!));
-//
-//                   return matchesAccount &&
-//                       matchesStatus &&
-//                       matchesSearch &&
-//                       matchesDate;
-//                 }).toList();
-//
-//                 if (orders.isEmpty) {
-//                   return const Center(
-//                     child: Text('No transactions available.'),
-//                   );
-//                 }
-//
-//                 return ListView.builder(
-//                   itemCount: orders.length,
-//                   itemBuilder: (context, index) {
-//                     final order = orders[index];
-//                     return OrderCard(order: order);
-//                   },
-//                 );
-//               },
-//             ),
-//           ),
-//         ],
-//       ),
-//     );
-//   }
-// }
-
-
 class TransactionScreen extends ConsumerWidget {
   final Accounts? account;
